TASK 47 - Refatorar modelos, serviços, etc. para facilitar manuseio no front (#27)
* remove all test files from shared

* adapt db for simplicity (please help)

* adapt calendar to use Task instead of ITask

* sidenav display relevant information for each page

* remove unused code in sidenav

* remove unused css, imports

* remove unused code

* ajustando banco aos novos modelos

---------

+++ src/app/layouts/memorybox-page/memorybox-page.component.ts
@@ -1,9 +1,11 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, ParamMap, Router} from "@angular/router";
 import {MemoryboxService} from "../../shared/services/memorybox.service";
-import {MemoryBox, Note, Task, User} from "../../shared/models";
+import {MemoryBox, Note, Tag, Task, User} from "../../shared/models";
 import {Observable} from 'rxjs'
 import {FormControl, Validators} from "@angular/forms";
+import {Imemorybox} from "../../shared/interfaces/imemorybox";
+import {CreateTagDialogComponent} from "../components/sidenav/create-tag-dialog/create-tag-dialog.component";
 @Component({
   selector: 'app-memorybox-page',
   templateUrl: './memorybox-page.component.html',
@@ -11,9 +13,10 @@ import {FormControl, Validators} from "@angular/forms";
 })
 export class MemoryboxPageComponent implements OnInit {
   public id: number = 0;
-  public memorybox: MemoryBox = new MemoryBox(0,"",0);
+  public memorybox!: MemoryBox;
   public isEditing: boolean = false;
   public titleForm: FormControl = new FormControl();
+  public isTagsCollapsed: boolean = false;
   constructor(private router: Router,
               private route: ActivatedRoute,
               private memoryBoxService: MemoryboxService) { }
@@ -36,16 +39,20 @@ export class MemoryboxPageComponent implements OnInit {
   }
 
   deleteMemoryBox() {
-    this.memoryBoxService.delete(this.memorybox.id).subscribe((res) => {
-      this.router.navigate(['/home']);
-    })
+    if (this.memorybox.id) {
+      this.memoryBoxService.delete(this.memorybox.id).subscribe((res) => {
+        this.router.navigate(['/home']);
+      })
+    }
   }
 
   editMemoryBox() {
-    this.memorybox.title = this.titleForm.value;
-    this.memoryBoxService.update(this.memorybox.id, this.memorybox).subscribe((res)=> {
-      this.toggleEditMemoryBox();
-    })
+    if (this.memorybox.id) {
+      this.memorybox.title = this.titleForm.value;
+      this.memoryBoxService.update(this.memorybox.id, this.memorybox).subscribe((res)=> {
+        this.toggleEditMemoryBox();
+      })
+    }
   }
 
   getErrorMessage() {
@@ -54,5 +61,19 @@ export class MemoryboxPageComponent implements OnInit {
     } else if (this.titleForm.hasError('minlength')) {
       return 'Anotação precisa ter no mínimo 4 carácteres!';
     } else { return ''}
+  }
+
+
+  toggleTags() {
+    this.isTagsCollapsed = !this.isTagsCollapsed;
+  }
+
+  deleteTag(tagARemover: number | undefined) {
+  }
+
+  editTag() {
+  }
+
+  openAddTagDialog() {
   }
 }
